% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestmatch.R
\name{best_match}
\alias{best_match}
\title{Matches the first level of geography to the second level of geography that it is the most nested in}
\usage{
best_match(overlap_output, id1)
}
\arguments{
\item{overlap_output}{The cleaned output from the overlapfunction.}

\item{id1}{The ID field of the first shapefile of interest}
}
\value{
The cleaned output with the first shapefile matched to the second level of geography for which it shares the most population
}
\description{
This function collapses the overlapfunction output by the first level of geography so that it can be mathed with the second level
of geography that it shares the most population.
}
\examples{
overlap115 <- overlapfunction(zip_cd115V, id1 =  "ZCTA5CE10", id2="GEOID", pop_field = "POP2010")
test4 <- best_match(overlap115, "ZCTA5CE10" )
}
